refactor: Update import paths for sidebar menu components
