+++ src/components/admin-panel/sidebar-menu/collapse-menu-button.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import Link from "next/link";
 import { useState } from "react";
 import { ChevronDown, Dot, LucideIcon } from "lucide-react";
